docs: include the README.md as main comment into the module lib.rs
