+++ crates/weaver_common/src/lib.rs
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
-//! A generic logger that can be used to log messages to the console.
+#![doc = include_str!("../README.md")]
 
 pub mod error;
 pub mod quiet;
