+++ crates/weaver_resolver/src/registry.rs
@@ -11,7 +11,7 @@ use weaver_resolved_schema::lineage::{AttributeLineage, GroupLineage};
 use weaver_resolved_schema::registry::{Constraint, Group, Registry};
 use weaver_semconv::attribute::AttributeSpec;
 use weaver_semconv::registry::SemConvRegistry;
-use weaver_semconv::GroupSpecWithProvenance;
+use weaver_semconv::group::GroupSpecWithProvenance;
 
 use crate::attribute::AttributeCatalog;
 use crate::constraint::resolve_constraints;
